python-pytest/pyproject.toml
[tool.poetry]
name = "python-pytest"
version = "0.1.0"
description = ""
authors = ["Roman Speranskii <r.speranskii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.3"
selene = {version = "^2.0.0-alpha.28", allow-prereleases = true}
selenium = "^3.141.0"
allure-pytest = "^2.8.16"

[tool.poetry.dev-dependencies]
pytest = "^4.6"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"